Create README.md
I developed it using CakePHP framework, utilizing composer for installation and cake bake for rapid development on backend. Database uses was PostgreSQL as this was my option for free database on Heroku.  The front-end was developed using Foundations CSS API and JQuery Javascript Framework for the logics behind and AJAX for server-client transactions for the remote control of the App.

Heroku Project Stage URL: https://connect-four-ariel.herokuapp.com/

By: Ariel Lim for Indinero